+++ .dockerignore
@@ -1,12 +1,6 @@
-.github/
-.git/
-docs/
-test/
-service/
-.travis.yml
-README.md
-LICENSE
-.gitattributes
-.gitignore
-.gitmodules
-.clang-format
+*
+!include/
+!src/
+!*.sh
+!CMakeLists.txt
+!vfrb.ini.in
