+++ .gitignore
@@ -6,10 +6,8 @@
 !/build/
 !/src/
 !/include/
-!/service/
 !/docs/
 !.github/**
-!/service/**
 !/docs/_includes/**
 !/docs/assets/**
 !LICENSE
@@ -18,14 +16,11 @@
 !.gitignore
 !.gitmodules
 !.clang-format
-!makefile
 !Dockerfile
 !.dockerignore
-!*.mk
 !*.sh
 !*.md
-!*.txt
-!*.ini
+!*.in
 !*.cpp
 !*.hpp
 !*.c
@@ -35,4 +30,6 @@
 !*.yaml
 !*.css
 !*.xml
-CMakeCache.txt
+!.gitkeep
+!CMakeLists.txt
+!version.txt
