--- service/vfrb.service.d/vfrb.conf
@@ -1,2 +1,0 @@
-[Service]
-Environment="LD_LIBRARY_PATH=%BOOST_LIBS_PATH%$LD_LIBRARY_PATH"
