+++ test/include/SocketImplTest.h
@@ -21,35 +21,36 @@
 
 #pragma once
 
-#include "server/NetworkInterface.hpp"
+#include <cstdint>
+#include <string>
+
+#include "util/defines.h"
 
 namespace server
 {
-class SocketImplTest;
-
-class NetworkInterfaceImplTests : public NetworkInterface<SocketImplTest>
+namespace net
+{
+class SocketImplTest
 {
 public:
-    NetworkInterfaceImplTests();
-    ~NetworkInterfaceImplTests() noexcept;
+    MOVABLE_BUT_NOT_COPYABLE(SocketImplTest)
 
-    void run(std::unique_lock<std::mutex>& lock) override;
+    explicit SocketImplTest(int&& socket);
+    ~SocketImplTest() noexcept;
 
-    void stop() override;
-
-    void onAccept(const std::function<void(bool)>& callback) override;
-
-    void close() override;
-
-    std::unique_ptr<Connection<SocketImplTest>> startConnection() override;
-
-    std::string get_currentAddress() const override;
-
-    void connect(bool err, const std::string& adr);
+    std::string get_address() const;
+    bool        write(const std::string& msg);
+    void        close();
+    int&        get();
 
 private:
-    std::function<void(bool)> on_accept;
-    std::string currentAddress;
-    bool stopped = false;
+    std::string  m_buffer;
+    std::int32_t m_socket;
+    std::string  m_address;
+
+public:
+    GETTER_CR(buffer)
+    SETTER_CR(address)
 };
+}  // namespace net
 }  // namespace server
